Moved TimeSeries methods into TimeSeriesDao.
